[Tizen][UI] Move TizenSystemIndicator to tizen/mobile/ui/
Move system indicator view and its helpers to tizen/mobile/ui/. The
rationale is that tizen/mobile/ directory stores components that are
self-contained and Tizen mobile specific.
