+++ tizen/mobile/ui/tizen_system_indicator_watcher.cc
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "xwalk/runtime/browser/ui/tizen_system_indicator_watcher.h"
+#include "xwalk/tizen/mobile/ui/tizen_system_indicator_watcher.h"
 
 #include <arpa/inet.h>
 #include <fcntl.h>
@@ -15,7 +15,7 @@
 #include "ipc/unix_domain_socket_util.h"
 #include "base/strings/string_tokenizer.h"
 #include "base/environment.h"
-#include "xwalk/runtime/browser/ui/tizen_system_indicator.h"
+#include "xwalk/tizen/mobile/ui/tizen_system_indicator.h"
 
 using content::BrowserThread;
